refact: adds generateSEO to dynamic work router
